ofz#4703 Direct-leak
Reviewed-on: https://gerrit.libreoffice.org/46893
Tested-by: Jenkins <ci@libreoffice.org>
Tested-by: Caolán McNamara <caolanm@redhat.com>

+++ starmath/source/rtfexport.cxx
@@ -412,7 +412,7 @@ void SmRtfExport::HandleBrace(const SmBraceNode* pNode, int nLevel)
     {
         auto body = static_cast<const SmBracebodyNode*>(pNode->Body());
         bool separatorWritten = false; // assume all separators are the same
-        for (int i = 0; i < body->GetNumSubNodes(); ++i)
+        for (size_t i = 0; i < body->GetNumSubNodes(); ++i)
         {
             const SmNode* subnode = body->GetSubNode(i);
             if (subnode->GetType() == SmNodeType::Math
